BGP: fix use of free memory by update_rsclient
BGP sometimes crashes when removing route server client because
of use after free.

The code to update rsclient created a local static copy of bgp attributes
but neglected to handle the extra information pointer. The extra information
was getting freed by bgp_attr_unintern() and reused later when the copy
was passed to bgp_attr_intern().

The fix is to use the attr_dup function to create a copy of
the extra information, then clean it up.

Signed-off-by: G.Balaji <balajig81@gmail.com>